fix: fix shop cost overflow
## 상점 오버플로우 수정
- 부품 강화 비용 계산을 long 타입으로 변경
- 직원 고용 비용 계산을 long 타입으로 변경
- 상점 UI 비용 비교 및 표시 타입을 long으로 통일
- 직원 슬롯 생산량 표시 계산을 long 타입으로 변경

+++ Assets/01.Scripts/Manager/EmployeeManager.cs
@@ -88,7 +88,7 @@ public class EmployeeManager : MonoBehaviour
         }
 
         // 4. 현재 고용 비용 계산
-        int hireCost = state.GetCurrentHireCost();
+        long hireCost = state.GetCurrentHireCost();
 
         // 5. CurrencyManager 존재 확인
         if (CurrencyManager.instance == null)
